Separate tokens to able to execute unittest.
