+++ API/line_api.py
@@ -1,20 +1,8 @@
-import json
-from dataclasses import dataclass
-
 import requests
 
+from API.tokens import line_tokens
 from crawlers.common import News
-from settings import TOKENS_JSON_PATH
 from utils import render_text_default
-
-
-@dataclass
-class Line:
-    channel_token: str
-
-
-with open(TOKENS_JSON_PATH, "r", encoding="utf-8") as rf:
-    line_tokens = Line(**json.load(rf)["line"])
 
 
 def broadcast(news: News, school_name: str) -> None:
